refactor: firebaseのconfigをファイル切り出し
